+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+# [1.42.0](https://github.com/adobe/helix-api-service/compare/v1.41.1...v1.42.0) (2026-04-07)
+
+
+### Features
+
+* **index:** add bulk indexing ([#172](https://github.com/adobe/helix-api-service/issues/172)) ([3a9f036](https://github.com/adobe/helix-api-service/commit/3a9f036c97a3bfbf9082c3dd0f7587d7ccbe8926))
+
+## [1.41.1](https://github.com/adobe/helix-api-service/compare/v1.41.0...v1.41.1) (2026-04-07)
+
+
+### Bug Fixes
+
+* **docs:** add git add for new api.html in website publish step ([#189](https://github.com/adobe/helix-api-service/issues/189)) ([23a3400](https://github.com/adobe/helix-api-service/commit/23a3400033083d7497499b93f7145fd551dc4b62))
+
+# [1.41.0](https://github.com/adobe/helix-api-service/compare/v1.40.1...v1.41.0) (2026-04-07)
+
+
+### Features
+
+* **docs:** add OpenAPI spec and generated API documentation ([#173](https://github.com/adobe/helix-api-service/issues/173)) ([3713efe](https://github.com/adobe/helix-api-service/commit/3713efe0fff5e2945f4c377d45d10fe955c187fa))
+
 ## [1.40.1](https://github.com/adobe/helix-api-service/compare/v1.40.0...v1.40.1) (2026-04-07)